@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
-import { Observable } from 'rxjs/Observable';
-import { ActivatedRoute } from '@angular/router';
+import { Observable } from 'rxjs';
+import { ActivatedRoute, Params } from '@angular/router';
+import { map } from 'rxjs/operators';
 
 @Component({
   selector: 'app-profile',
@@ -11,7 +12,8 @@ export class ProfileComponent {
   user: Observable<string>;
 
   constructor(private route: ActivatedRoute) {
-    this.user = route.params.map((p: any) => p.id);
+    this.user = this.route.params
+      .pipe(map((p: Params) => p.id as string));
   }
 
 }
