@@ -1,9 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Member } from './';
-import { Observable } from 'rxjs/Observable';
+import { Observable, of, forkJoin } from 'rxjs';
 import { HttpClient } from '@angular/common/http';
-import 'rxjs/add/observable/forkJoin';
-import 'rxjs/add/observable/of';
 
 @Injectable()
 export class MembersService {
@@ -15,7 +13,7 @@ export class MembersService {
   constructor(private http: HttpClient) {
     const observables = this.usernames.map((username: string) => this.getMemberByHandle(username));
 
-    this.members$ = Observable.forkJoin(...observables);
+    this.members$ = forkJoin(...observables);
     this.members$.subscribe();
   }
 
@@ -25,7 +23,7 @@ export class MembersService {
 
   getMemberByHandle(username): Observable<Member> {
     if (username === 'none') {
-      return Observable.of(void 0);
+      return of(void 0);
     }
     return this.http.get<Member>(`${this.gitHubUrl}/${username}`);
   }
