@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { LocalizeRouterService } from 'localize-router';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-root',
@@ -9,7 +10,9 @@ import { LocalizeRouterService } from 'localize-router';
 export class AppComponent {
   title = 'app';
 
-  constructor(private localize: LocalizeRouterService) {
+  constructor(private localize: LocalizeRouterService, private router: Router) {
+    console.log('PARSER ROUTES', this.localize.parser.routes);
+    console.log('ROUTER ROUTES', this.router.config);
   }
 
   changeLanguage(lang: string) {
